BUG: Changed a few types from int or unsigned int to size_t. This makes the casts that were used unnecassary. This improves 64bit compatibility on LP64 systems

+++ Common/Testing/Cxx/vtkTestUtilities.h
@@ -150,7 +150,7 @@ char* vtkTestUtilities::ExpandFileNameWithArgOrEnvOrDefault(const char* arg,
     fullName = new char[strlen(value) + strlen(fname) + 2 + (slash ? 1 : 0)];
     fullName[0] = 0;
     strcat(fullName, value);
-    int len = static_cast<int>(strlen(fullName));
+    size_t len = strlen(fullName);
     fullName[len] = '/';
     fullName[len+1] = 0;
     strcat(fullName, fname);
